usage/requirements.txt
tensorflow-gpu
soundfile